+++ src/converters/fromZigbee.ts
@@ -1980,102 +1980,6 @@ export const hw_version: Fz.Converter<"genBasic", undefined, ["attributeReport",
 // #endregion
 
 // #region Non-generic converters
-export const ptvo_switch_uart: Fz.Converter<"genMultistateValue", undefined, ["attributeReport", "readResponse"]> = {
-    cluster: "genMultistateValue",
-    type: ["attributeReport", "readResponse"],
-    convert: (model, msg, publish, options, meta) => {
-        let data: unknown[] | string = msg.data.stateText as unknown[]; // ZclArray is only for write
-        if (Array.isArray(data)) {
-            let bHex = false;
-            let code: number;
-            let index: number;
-            for (index = 0; index < data.length; index += 1) {
-                code = data[index] as number;
-                if (code < 32 || code > 127) {
-                    bHex = true;
-                    break;
-                }
-            }
-            if (!bHex) {
-                data = data.toString();
-            } else {
-                data = [...data];
-            }
-        }
-        return {action: data};
-    },
-};
-export const ptvo_switch_analog_input: Fz.Converter<"genAnalogInput", undefined, ["attributeReport", "readResponse"]> = {
-    cluster: "genAnalogInput",
-    type: ["attributeReport", "readResponse"],
-    convert: (model, msg, publish, options, meta) => {
-        const payload: KeyValueAny = {};
-        const channel = msg.endpoint.ID;
-        const name = `l${channel}`;
-        const endpoint = msg.endpoint;
-        payload[name] = precisionRound(msg.data.presentValue, 3);
-        const cluster = "genLevelCtrl";
-        if (endpoint && (endpoint.supportsInputCluster(cluster) || endpoint.supportsOutputCluster(cluster))) {
-            payload[`brightness_${name}`] = msg.data.presentValue;
-        } else if (msg.data.description !== undefined) {
-            const data1 = msg.data.description;
-            if (data1) {
-                const data2 = data1.split(",");
-                const devid = data2[1];
-                const unit = data2[0];
-                if (devid) {
-                    payload[`device_${name}`] = devid;
-                }
-
-                const valRaw = msg.data.presentValue;
-                if (unit) {
-                    let val = precisionRound(valRaw, 1);
-
-                    const nameLookup: KeyValueAny = {
-                        C: "temperature",
-                        "%": "humidity",
-                        m: "altitude",
-                        Pa: "pressure",
-                        ppm: "quality",
-                        psize: "particle_size",
-                        V: "voltage",
-                        A: "current",
-                        Wh: "energy",
-                        W: "power",
-                        Hz: "frequency",
-                        pf: "power_factor",
-                        lx: "illuminance",
-                    };
-
-                    let nameAlt = "";
-                    if (unit === "A" || unit === "pf") {
-                        if (valRaw < 1) {
-                            val = precisionRound(valRaw, 3);
-                        }
-                    }
-                    if (unit.startsWith("mcpm") || unit.startsWith("ncpm")) {
-                        const num = unit.substr(4, 1);
-                        nameAlt = num === "A" ? `${unit.substr(0, 4)}10` : unit;
-                        val = precisionRound(valRaw, 2);
-                    } else {
-                        nameAlt = nameLookup[unit];
-                    }
-                    if (nameAlt === undefined) {
-                        const valueIndex = Number.parseInt(unit, 10);
-                        if (!Number.isNaN(valueIndex)) {
-                            nameAlt = `val${unit}`;
-                        }
-                    }
-
-                    if (nameAlt !== undefined) {
-                        payload[`${nameAlt}_${name}`] = val;
-                    }
-                }
-            }
-        }
-        return payload;
-    },
-};
 export const tint_scene: Fz.Converter<"genBasic", undefined, "write"> = {
     cluster: "genBasic",
     type: "write",
@@ -2145,16 +2049,6 @@ export const command_status_change_notification_action: Fz.Converter<"ssIasZone"
         return {action: lookup[msg.data.zonestatus]};
     },
 };
-export const ptvo_multistate_action: Fz.Converter<"genMultistateInput", undefined, ["attributeReport", "readResponse"]> = {
-    cluster: "genMultistateInput",
-    type: ["attributeReport", "readResponse"],
-    convert: (model, msg, publish, options, meta) => {
-        const actionLookup: KeyValueAny = {0: "release", 1: "single", 2: "double", 3: "tripple", 4: "hold"};
-        const value = msg.data.presentValue;
-        const action = actionLookup[value];
-        return {action: postfixWithEndpointName(action, msg, model, meta)};
-    },
-};
 // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
 export const U02I007C01_contact: Fz.Converter<"ssIasZone", undefined, "commandStatusChangeNotification"> = {
     cluster: "ssIasZone",
@@ -2177,51 +2071,6 @@ export const U02I007C01_water_leak: Fz.Converter<"ssIasZone", undefined, "comman
         return {
             water_leak: (zoneStatus & 1) > 0,
         };
-    },
-};
-export const itcmdr_clicks: Fz.Converter<"genMultistateInput", undefined, ["readResponse", "attributeReport"]> = {
-    cluster: "genMultistateInput",
-    type: ["readResponse", "attributeReport"],
-    convert: (model, msg, publish, options, meta) => {
-        const lookup: KeyValueAny = {0: "hold", 1: "single", 2: "double", 3: "triple", 4: "quadruple", 255: "release"};
-        const clicks = msg.data.presentValue;
-        const action = lookup[clicks] ? lookup[clicks] : "many";
-        return {action};
-    },
-};
-// biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
-export const ZB003X_attr: Fz.Converter<"ssIasZone", undefined, ["attributeReport", "readResponse"]> = {
-    cluster: "ssIasZone",
-    type: ["attributeReport", "readResponse"],
-    convert: (model, msg, publish, options, meta) => {
-        const data = msg.data;
-        const senslookup: Record<number, string> = {0: "low", 1: "medium", 2: "high"};
-        const keeptimelookup: Record<number, number> = {0: 0, 1: 30, 2: 60, 3: 120, 4: 240, 5: 480};
-        if (data && data.currentZoneSensitivityLevel !== undefined) {
-            const value = data.currentZoneSensitivityLevel;
-            return {sensitivity: senslookup[value]};
-        }
-        if (data && data["61441"] !== undefined) {
-            const value = data["61441"] as number;
-            return {keep_time: keeptimelookup[value]};
-        }
-    },
-};
-// biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
-export const ZB003X_occupancy: Fz.Converter<"ssIasZone", undefined, "commandStatusChangeNotification"> = {
-    cluster: "ssIasZone",
-    type: "commandStatusChangeNotification",
-    convert: (model, msg, publish, options, meta) => {
-        const zoneStatus = msg.data.zonestatus;
-        return {occupancy: (zoneStatus & 1) > 0, tamper: (zoneStatus & 4) > 0};
-    },
-};
-export const rc_110_level_to_scene: Fz.Converter<"genLevelCtrl", undefined, ["commandMoveToLevel", "commandMoveToLevelWithOnOff"]> = {
-    cluster: "genLevelCtrl",
-    type: ["commandMoveToLevel", "commandMoveToLevelWithOnOff"],
-    convert: (model, msg, publish, options, meta) => {
-        const scenes: KeyValueAny = {2: "1", 52: "2", 102: "3", 153: "4", 194: "5", 254: "6"};
-        return {action: `scene_${scenes[msg.data.level]}`};
     },
 };
 const SUNRICHER_SWITCH2801K2_LOOKUP: Record<number, string> = {
@@ -2510,16 +2359,6 @@ export const command_arm_with_transaction: Fz.Converter<"ssIasAce", undefined, "
         if (!payload) return;
         payload.action_transaction = msg.meta.zclTransactionSequenceNumber;
         return payload;
-    },
-};
-export const command_on_presence: Fz.Converter<"genOnOff", undefined, "commandOn"> = {
-    cluster: "genOnOff",
-    type: "commandOn",
-    convert: (model, msg, publish, options, meta) => {
-        const newMsg = {...msg, type: "commandCheckin" as const, data: {}};
-        const payload1 = checkin_presence.convert(model, newMsg, publish, options, meta);
-        const payload2 = command_on.convert(model, msg, publish, options, meta);
-        return {...payload1, ...payload2};
     },
 };
 export const ias_ace_occupancy_with_timeout: Fz.Converter<"ssIasAce", undefined, "commandGetPanelStatus"> = {